backend/requirements.txt
alembic==1.13.1
attrs==24.3.0
beautifulsoup4==4.12.2
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
deprecated==1.2.15
flask==2.3.3
flask-cors==4.0.0
flask-limiter==3.5.0
flask-migrate==4.0.5
flask-sqlalchemy==3.1.1
flask-testing==0.8.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
jinja2==3.1.5
limits==3.14.1
mako==1.3.8
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.7
pygments==2.18.0
pysocks==1.7.1
pytest==7.4.2
python-dotenv==1.0.0
requests==2.31.0
rich==13.9.4
schedule==1.2.2
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
splinter==0.21.0
sqlalchemy==2.0.21
trio==0.28.0
trio-websocket==0.11.1
typing-extensions==4.12.2
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
werkzeug==3.1.3
wrapt==1.17.0
wsproto==1.2.0
